Silence compiler warnings
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@11491 8aca7d54-2c30-db11-9de9-000461428c89

+++ src/simutrans/tpl/freelist_tpl.h
@@ -100,7 +100,7 @@ public:
 			chunk_list = chunk;
 			p += sizeof(nodelist_node_t);
 			// then enter nodes into nodelist
-			for (int i = 0; i < new_chuck_size; i++) {
+			for (size_t i = 0; i < new_chuck_size; i++) {
 				nodelist_node_t* tmp = (nodelist_node_t*)(p + i*NODE_SIZE);
 #ifdef USE_VALGRIND_MEMCHECK
 				// tell valgrind that we reserved space for one nodelist_node_t
